+++ include/openrave/openrave.h
@@ -1167,6 +1167,14 @@ public:
         }
     }
 
+    void Reset() {
+        _transform.identity();
+        _type = IKP_None;
+        _id.clear();
+        _name.clear();
+        _mapCustomData.clear();
+    }
+
     inline IkParameterizationType GetType() const {
         return _type;
     }